+++ src/compas/geometry/curves/nurbs.py
@@ -123,6 +123,11 @@ class NurbsCurve(Curve):
     # ==============================================================================
 
     @property
+    def dtype(self):
+        """str : The type of the object in the form of a '2-level' import and a class name."""
+        return 'compas.geometry/NurbsCurve'
+
+    @property
     def data(self):
         return {
             'points': [point.data for point in self.points],