+++ src/compas_blender/artists/robotmodelartist.py
@@ -7,6 +7,7 @@ import bpy
 import mathutils
 
 import compas_blender
+from compas_blender.utilities import RGBColor
 from compas.datastructures import Mesh
 from compas.geometry import Transformation, Shape
 from compas.robots import RobotModel
@@ -37,7 +38,7 @@ class RobotModelArtist(BlenderArtist, RobotModelArtist):
     def create_geometry(self,
                         geometry: Union[Mesh, Shape],
                         name: str = None,
-                        color: Union[Tuple[int, int, int, int], Tuple[float, float, float, float]] = None
+                        color: Union[RGBColor, Tuple[int, int, int, int], Tuple[float, float, float, float]] = None
                         ) -> bpy.types.Object:
         # Imported colors take priority over a the parameter color
         if 'mesh_color.diffuse' in geometry.attributes:
@@ -45,8 +46,7 @@ class RobotModelArtist(BlenderArtist, RobotModelArtist):
 
         # If we have a color, we'll discard alpha because draw_mesh is hard coded for a=1
         if color:
-            r, g, b, _a = color
-            color = (r, g, b)
+            color = color[:3]
         else:
             color = (1., 1., 1.)