+++ CHANGELOG.md
@@ -9,7 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `CircleArtist`, `LineArtist`, `PointArtist`, `PolygonArtist`, `PolylineArtist`, and `VectorArtist` to `compas_blender`.
+* Added `draw_circles` and `draw_planes` to `compas_blender`.
+* Added `compas_rhino.geometry.curves` plugins for `compas.geometry.curves` pluggables.
+* Added `compas_rhino.geometry.RhinoNurbsCurve`.
+
 ### Changed
+
+* Replaced implementation of `RGBColour` and `Float` with deprecation warning in `compas.utilities.descriptors`.
+* Moved all Rhino geometry and objects wrappers to `compas_rhino.conversions`.
+* Fixed bug in `compas_rhino.conversions.RhinoSurface.from_geometry`.
+* Changed `compas_rhino.conversions.RhinoLine.from_geometry` to accept line curves.
+* Fixed bug in `compas_rhino.geometry.RhinoNurbsCurve.closest_point`.
 
 ### Removed